Add basic string interpolation for date and time in tags
This adds string interpolation for date/time in tags, so variable
fields `{date}` and `{time}` can be used to generate semi-unique image
tags via the CLI and build.yml. The variables available for
interpolation may be expanded on in the future.

As the resulting tags aren't necessarily known in advance, dbuild now
also prints a summary of generated artifacts when finished.